WIP: Convert files to typescript. Tests do not pass.
