+++ .gitignore
@@ -8,6 +8,9 @@ node_modules
 build
 dist
 
+# typescript builds
+.rpt2_cache
+
 # test coverage
 coverage
 
@@ -22,3 +25,4 @@ coverage
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
