+++ rollup.config.js
@@ -4,11 +4,12 @@ import external from "rollup-plugin-peer-deps-external";
 import postcss from "rollup-plugin-postcss";
 import resolve from "rollup-plugin-node-resolve";
 import url from "rollup-plugin-url";
+import typescript from "rollup-plugin-typescript2";
 
 import pkg from "./package.json";
 
 export default {
-  input: "src/index.js",
+  input: "src/index.tsx",
   output: [
     {
       file: pkg.main,
@@ -28,6 +29,7 @@ export default {
       modules: true
     }),
     url(),
+    typescript(),
     babel({
       exclude: "node_modules/**"
     }),
